requirements.txt
gradio
openai
numpy
fastapi
uvicorn
python-dotenv
html5lib